Another big migration point.
What went wrong? Well aparently theres a thing called a cascade delete issue. Basiclly Users have Recpies and Favorites. Users post recipies, and can list recipies as their "favorite" (either their own or others)

Favorite has an FK for both Users and Recipies, so if you delete an user, or a recipe SQL doesnt like that apparenntly. This is because it causes "mutliple cascade paths" where the database is in a loop of trying to delete things that don't exist (i think?) so i have to sort of path it out for them. wierd

Basically If I delete a User then their Favorites get delelted
and If I delete a Recipe, all favorites attached to it get deleted. (Like if you save a Youtube video, if your account gets deleted then the data about your saved videos go away, and if the video itself is deleted then it can't be saved anywhere)

Now, Im not sure Why this was only a problem with Favorites and not Reviews because they seem to operate in a similar manner, atleast thast was my inntetion.

Anyways my solutuin is to either have a single delete path which is perfectly OK or to just null out the FK value meaning the remaining reviews dont get broken by have "no FK" but rather a NULL one so it doesnt break a connectted properties. With one path setting null values the other goes on a SOLO cascade path, a cleanup if you will. I really really hope it works

this is probably horribly inefficient and theres a much simplilar solution but im this far so i whatever